Revert "DecodeRawTransaction: Change locktime type to long"
This reverts commit 37d3b45d7289aceee95b9ebb814199d35374bdc0.

+++ bitcoind/src/main/java/bisq/wallets/bitcoind/rpc/responses/AbstractDecodeRawTransactionResponse.java
@@ -33,7 +33,7 @@ public class AbstractDecodeRawTransactionResponse<S extends AbstractVin, T exten
     private int version;
 
     @Json(name = "locktime")
-    private long lockTime;
+    private int lockTime;
     private List<S> vin;
     private List<T> vout;
 }
